src/pages/functions/requirements.txt
pandas
numpy
scikit-learn
joblib
google-cloud-firestore
firebase-admin
